pdate env variable part 36

+++ pages/api/products/[id].ts
@@ -18,16 +18,22 @@ const handlers = async (req: NextApiRequest, res: NextApiResponse) => {
           return getOneProduct(id as string);
         }
       case "POST":
-        return createProduct(req, res);
+        if (session?.role === "admin") {
+          return createProduct(req, res);
+        }
       case "PUT":
-        return editProduct(req, res);
+        if (session?.role === "admin") {
+          return editProduct(req, res);
+        }
       case "DELETE":
-        return deleteProduct(req, res);
+        if (session?.role === "admin") {
+          return deleteProduct(req, res);
+        }
       default:
         return res.status(405).end(`Method ${req.method} Not Allowed`);
     }
   } catch (e) {
-    console.log("ERRRRRORRRRRRRRRR", e);
+    console.log(e);
   }
 };
